refactor: seo component to proper use data
